#1686 - Start Refactor 3, change storage type to long (bigint), use jsonb for most stat data, store game data in tables. Change some id's to snowflakes.

+++ core/src/main/java/me/mykindos/betterpvp/core/client/achievements/impl/game/dom/capture/PointsCapturedAchievement.java
@@ -29,7 +29,7 @@ public class PointsCapturedAchievement extends SingleSimpleAchievement {
         super("Points Captured", key,
                 AchievementCategories.GAME_POINTS_CAPTURED,
                 AchievementType.GLOBAL,
-                (double) goal,
+                (long) goal,
                 GameTeamMapNativeStat.builder()
                         .action(GameTeamMapNativeStat.Action.CONTROL_POINT_CAPTURED)
                         .build()
